fix compile and remove warnings

+++ contracts/utils/ERC721Test.sol
@@ -1,12 +1,14 @@
 // SPDX-License-Identifier: MIT
 pragma solidity 0.8.10;
 
-import "@openzeppelin/contracts/token/ERC721/ERC721.sol";
+import '@openzeppelin/contracts/token/ERC721/ERC721.sol';
 
 contract ERC721Test is ERC721 {
-    uint tokenId = 5;
+    uint256 tokenId = 5;
 
-    constructor(string memory _name, string memory _symbol) ERC721(_name, _symbol) {
+    constructor(string memory _name, string memory _symbol)
+        ERC721(_name, _symbol)
+    {
         _safeMint(msg.sender, tokenId);
     }
 }
